record batch fixes
