+++ protocol/fetch/fetch.go
@@ -2,7 +2,6 @@ package fetch
 
 import (
 	"fmt"
-	"io"
 
 	"github.com/segmentio/kafka-go/protocol"
 )
@@ -94,7 +93,7 @@ func (r *Response) Close() error {
 
 		for j := range t.Partitions {
 			p := &t.Partitions[j]
-			closeRecordBatch(p.RecordSet.Records)
+			protocol.CloseRecordBatch(p.RecordSet.Records)
 		}
 	}
 
@@ -107,20 +106,8 @@ func (r *Response) Reset() {
 
 		for j := range t.Partitions {
 			p := &t.Partitions[j]
-			resetRecordBatch(p.RecordSet.Records)
+			protocol.ResetRecordBatch(p.RecordSet.Records)
 		}
-	}
-}
-
-func closeRecordBatch(rb protocol.RecordBatch) {
-	if r, _ := rb.(io.Closer); r != nil {
-		r.Close()
-	}
-}
-
-func resetRecordBatch(rb protocol.RecordBatch) {
-	if r, _ := rb.(interface{ Reset() }); r != nil {
-		r.Reset()
 	}
 }
