+++ time.go
@@ -12,7 +12,7 @@ const (
 )
 
 func makeTime(t int64) time.Time {
-	return time.Unix(t/1000, (t%1000)*int64(time.Millisecond))
+	return time.Unix(t/1000, (t%1000)*int64(time.Millisecond)).UTC()
 }
 
 func timestamp(t time.Time) int64 {
